Don't warn with HWLOC API version 3
Leading up to major version 3.0, hwloc removed the configuration
option for 16-bit domains (their commit bf44f300e87). This can cause a
compiler warning when GROMACS tries to store the new int value into
our 16-bit field. Since we don't actually care how wide the field is,
make it 32 bits and it will always work without warnings.

+++ src/gromacs/hardware/hardwaretopology.h
@@ -186,7 +186,7 @@ public:
         std::uint16_t vendorId;   //!< Vendor identification
         std::uint16_t deviceId;   //!< Vendor-specific device identification
         std::uint16_t classId;    //!< class (high 8 bits) and subclass (low 8 bits)
-        std::uint16_t domain;     //!< Domain, usually 0 for PCI bus
+        std::uint32_t domain;     //!< Domain, usually 0 for PCI bus
         std::uint8_t  bus;        //!< Bus number in domain
         std::uint8_t  dev;        //!< Device on bus
         std::uint8_t  func;       //!< Function id for multi-function devices
